requirements.txt
pyside6
sqlalchemy
python-dotenv
asyncpg
pandas
openpyxl
ruff